rumi

+++ dev/application/views/Admin/showManagePageSec.php
@@ -30,7 +30,7 @@
                 <div class="btn-group">
 
                     <a class="btn" href="<?php echo base_url()?>Admin/Page/editPageSectionShow/<?php echo $pg->pageSectionId?>"><i class="icon_pencil-edit"></i></a>
-                    <a class="btn " href="<?php echo base_url()?>Admin/Page/deletePageSection/<?php echo $pg->pageSectionId?>"><i class="icon_trash"></i></a>
+                    <a class="btn " href="<?php echo base_url()?>Admin/Page/deletePageSection/<?php echo $pg->pageSectionId?>" onclick='return confirm("Are you sure?")'><i class="icon_trash"></i></a>
 
                 </div>
             </td>
@@ -42,24 +42,3 @@
 
     </tbody>
 </table>
-<script>
-    function selectid(x) {
-        if (confirm('Are you sure you want to delete this Page Section !! ?')) {
-            btn = $(x).data('panel-id');
-            $.ajax({
-                type:'POST',
-                url:'<?php echo base_url("Admin/Page/deletePageSection/")?>'+btn,
-                data:{'pageId':btn},
-                cache: false,
-                success:function(data) {
-                    alert("Page Section Deleted Successfully!!");
-                    location.reload();
-                }
-            });
-        }
-        else {
-
-            location.reload();
-        }
-    }
-</script>
